Refactored mainwindow to MainWindow to fit our CamelCase convention #195
