+++ gui/src/main.cpp
@@ -14,7 +14,7 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 ********************************************************************************/
-#include "mainwindow.h"
+#include "MainWindow.h"
 #include <QApplication>
 
 int main(int argc, char *argv[]){
